cleaned up and fixed tests
