+++ semantic_digital_twin/src/semantic_digital_twin/reasoning/world_reasoner.py
@@ -56,7 +56,8 @@ class WorldReasoner:
             != self._last_world_model_version
         ):
             self.reasoner.result = self.reasoner.rdr.classify(self.world)
-            self._update_world_attributes()
+            with self.world.modify_world():
+                self._update_world_attributes()
             self._last_world_model_version = (
                 self.world.get_world_model_manager().version
             )
