+++ test/semantic_digital_twin_test/test_worlds/test_reasoning_queries.py
@@ -1,7 +1,18 @@
-from semantic_digital_twin.predetermined_maps.kitchen_environment import KitchenEnvironment
-from semantic_digital_twin.reasoning.queries import semantic_annotations_on_surfaces, \
-    get_next_object_using_planar_distance, goal_surface_of_object, filter_annotations_by_color, \
-    annotation_class_by_label, sort_annotations_by_volume
+import os
+
+from semantic_digital_twin.adapters.urdf import URDFParser
+from semantic_digital_twin.predetermined_maps.kitchen_environment import (
+    KitchenEnvironment,
+)
+from semantic_digital_twin.reasoning.queries import (
+    semantic_annotations_on_surfaces,
+    get_next_object_using_planar_distance,
+    goal_surface_of_object,
+    filter_annotations_by_color,
+    annotation_class_by_label,
+    sort_annotations_by_volume,
+)
+from semantic_digital_twin.reasoning.world_reasoner import WorldReasoner
 from semantic_digital_twin.semantic_annotations.semantic_annotations import *
 from semantic_digital_twin.world import World
 from semantic_digital_twin.world_description.geometry import Color
@@ -16,12 +27,33 @@ def test_load_environment_returns_world():
     assert world.root.name == PrefixedName("root")
 
 
+def test_world_reasoner_reason_returns_dicts():
+    urdf_dir = os.path.join(
+        os.path.dirname(os.path.abspath(__file__)),
+        "..",
+        "..",
+        "semantic_digital_twin",
+        "resources",
+        "urdf",
+    )
+    world = URDFParser.from_file(
+        file_path=os.path.join(urdf_dir, "kitchen-small.urdf")
+    ).parse()
+    reasoner = WorldReasoner(world)
+
+    result = reasoner.reason()
+
+    assert isinstance(result, dict)
+
+
 def test_semantic_annotations_on_surfaces(kitchen_environment_fixture):
     """
     Tests that giving Table annotations gives a list of the correct annotation on top.
     """
     table1 = kitchen_environment_fixture.get_semantic_annotation_by_name("fruit_table")
-    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name("vegetable_table")
+    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name(
+        "vegetable_table"
+    )
     table3 = kitchen_environment_fixture.get_semantic_annotation_by_name("empty_table")
     apple = kitchen_environment_fixture.get_semantic_annotation_by_name("apple")
     carrot = kitchen_environment_fixture.get_semantic_annotation_by_name("carrot")
@@ -29,13 +61,9 @@ def test_semantic_annotations_on_surfaces(kitchen_environment_fixture):
     lettuce = kitchen_environment_fixture.get_semantic_annotation_by_name("lettuce")
     banana1 = kitchen_environment_fixture.get_semantic_annotation_by_name("banana1")
 
-    assert semantic_annotations_on_surfaces([table1, table2, table3], kitchen_environment_fixture) == [
-        apple,
-        orange,
-        banana1,
-        carrot,
-        lettuce
-    ]
+    assert semantic_annotations_on_surfaces(
+        [table1, table2, table3], kitchen_environment_fixture
+    ) == [apple, orange, banana1, carrot, lettuce]
     assert semantic_annotations_on_surfaces([], kitchen_environment_fixture) == []
 
 
@@ -50,7 +78,9 @@ def test_get_next_object_using_planar_distance(kitchen_environment_fixture):
     """
     toya = kitchen_environment_fixture.get_body_by_name("base_link_body")
     table1 = kitchen_environment_fixture.get_semantic_annotation_by_name("fruit_table")
-    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name("vegetable_table")
+    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name(
+        "vegetable_table"
+    )
     table3 = kitchen_environment_fixture.get_semantic_annotation_by_name("empty_table")
     apple = kitchen_environment_fixture.get_semantic_annotation_by_name("apple")
     carrot = kitchen_environment_fixture.get_semantic_annotation_by_name("carrot")
@@ -58,15 +88,25 @@ def test_get_next_object_using_planar_distance(kitchen_environment_fixture):
     lettuce = kitchen_environment_fixture.get_semantic_annotation_by_name("lettuce")
     banana1 = kitchen_environment_fixture.get_semantic_annotation_by_name("banana1")
 
-    assert get_next_object_using_planar_distance(toya, table1, Vector3.Z()).tolist() == [orange, banana1, apple]
-    assert get_next_object_using_planar_distance(toya, table2, Vector3.Z()).tolist() == [
+    assert get_next_object_using_planar_distance(
+        toya, table1, Vector3.Z()
+    ).tolist() == [orange, banana1, apple]
+    assert get_next_object_using_planar_distance(
+        toya, table2, Vector3.Z()
+    ).tolist() == [
         carrot,
         lettuce,
     ]
-    assert get_next_object_using_planar_distance(apple.bodies[0], table1, Vector3.Z()).tolist() == [apple, banana1, orange]
-    assert get_next_object_using_planar_distance(apple.bodies[0], table1, Vector3.Y()).tolist() == [apple, orange, banana1]
+    assert get_next_object_using_planar_distance(
+        apple.bodies[0], table1, Vector3.Z()
+    ).tolist() == [apple, banana1, orange]
+    assert get_next_object_using_planar_distance(
+        apple.bodies[0], table1, Vector3.Y()
+    ).tolist() == [apple, orange, banana1]
 
-    assert get_next_object_using_planar_distance(toya, table3, Vector3.Z()).tolist() == []
+    assert (
+        get_next_object_using_planar_distance(toya, table3, Vector3.Z()).tolist() == []
+    )
 
 
 def test_goal_surface_of_object(kitchen_environment_fixture):
@@ -78,7 +118,9 @@ def test_goal_surface_of_object(kitchen_environment_fixture):
     no valid candidates.
     """
     table1 = kitchen_environment_fixture.get_semantic_annotation_by_name("fruit_table")
-    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name("vegetable_table")
+    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name(
+        "vegetable_table"
+    )
     table3 = kitchen_environment_fixture.get_semantic_annotation_by_name("empty_table")
     table4 = kitchen_environment_fixture.get_semantic_annotation_by_name("empty_table2")
 
@@ -103,6 +145,7 @@ def test_goal_surface_of_object(kitchen_environment_fixture):
     assert goal_surface_of_object(apple, [table2, table3, table4]) == table3
     assert goal_surface_of_object(apple, [table2, table4, table3]) == table4
 
+
 def test_filter_annotations_by_color(kitchen_environment_fixture):
     """
     Tests the filter_annotations_by_color function by verifying the retrieval of semantic
@@ -113,16 +156,31 @@ def test_filter_annotations_by_color(kitchen_environment_fixture):
     within the given list of objects.
     """
     table1 = kitchen_environment_fixture.get_semantic_annotation_by_name("fruit_table")
-    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name("vegetable_table")
+    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name(
+        "vegetable_table"
+    )
     apple = kitchen_environment_fixture.get_semantic_annotation_by_name("apple")
     orange = kitchen_environment_fixture.get_semantic_annotation_by_name("orange")
     carrot = kitchen_environment_fixture.get_semantic_annotation_by_name("carrot")
 
     assert filter_annotations_by_color(Color.RED(), [apple, orange]).tolist() == [apple]
-    assert filter_annotations_by_color(Color.ORANGE(), [apple, orange]).tolist() == [orange]
-    assert filter_annotations_by_color(Color.BLUE(), [apple, orange,carrot]).tolist() == []
-    assert filter_annotations_by_color(Color.ORANGE(), (semantic_annotations_on_surfaces([table1, table2], kitchen_environment_fixture))).tolist() == [orange, carrot]
+    assert filter_annotations_by_color(Color.ORANGE(), [apple, orange]).tolist() == [
+        orange
+    ]
+    assert (
+        filter_annotations_by_color(Color.BLUE(), [apple, orange, carrot]).tolist()
+        == []
+    )
+    assert filter_annotations_by_color(
+        Color.ORANGE(),
+        (
+            semantic_annotations_on_surfaces(
+                [table1, table2], kitchen_environment_fixture
+            )
+        ),
+    ).tolist() == [orange, carrot]
     assert filter_annotations_by_color(Color.YELLOW(), []).tolist() == []
+
 
 def test_annotation_class_by_label():
     """
@@ -140,14 +198,26 @@ def test_sort_annotations_by_volume(kitchen_environment_fixture):
     Tests the sort_annotations_by_volume function by verifying the order of the returned annotations.
     """
     table1 = kitchen_environment_fixture.get_semantic_annotation_by_name("fruit_table")
-    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name("vegetable_table")
+    table2 = kitchen_environment_fixture.get_semantic_annotation_by_name(
+        "vegetable_table"
+    )
 
     apple = kitchen_environment_fixture.get_semantic_annotation_by_name("apple")
     carrot = kitchen_environment_fixture.get_semantic_annotation_by_name("carrot")
     lettuce = kitchen_environment_fixture.get_semantic_annotation_by_name("lettuce")
 
-    assert sort_annotations_by_volume([]) ==[]
+    assert sort_annotations_by_volume([]) == []
     assert sort_annotations_by_volume([apple, carrot]) == [apple, carrot]
-    assert sort_annotations_by_volume([table1, lettuce, apple]) == [table1, lettuce, apple]
-    assert sort_annotations_by_volume([table1, lettuce, apple], False) == [apple, lettuce, table1]
-    assert sort_annotations_by_volume(semantic_annotations_on_surfaces([table2], kitchen_environment_fixture)) == [lettuce, carrot]
+    assert sort_annotations_by_volume([table1, lettuce, apple]) == [
+        table1,
+        lettuce,
+        apple,
+    ]
+    assert sort_annotations_by_volume([table1, lettuce, apple], False) == [
+        apple,
+        lettuce,
+        table1,
+    ]
+    assert sort_annotations_by_volume(
+        semantic_annotations_on_surfaces([table2], kitchen_environment_fixture)
+    ) == [lettuce, carrot]
